+++ CHANGELOG.md
@@ -10,6 +10,12 @@
   - update Docker RUNTIME_BASE=alpine:3.21
   - remove unused dependencies
   - remove swagger
+- [TD-5713]
+  - Updates analyzers to take `asciifolding` and multiple languages into account.
+  - `search_as_you_type` type in concept name.
+  - Defines scope for native and dynamic fields search.
+  - Use query data in `ElasticDocumentProtocol` to build queries.
+
 
 ## [6.16.0] 2024-12-16